pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "servingbench-tuner"
version = "0.1.0"
description = "ServingBench-Tuner: production-oriented LLM inference serving benchmark + quality gate + multi-objective tuner + reproducible regression"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "inference", "benchmark", "serving", "nsga2", "tuning", "sla", "reproducibility"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Software Development :: Testing",
  "Topic :: System :: Benchmark",
]

# Base dependencies: keep it portable (no torch/vllm in base)
dependencies = [
  "pydantic>=2.6,<3",
  "PyYAML>=6.0.1,<7",
  "typer>=0.12,<1",
  "rich>=13.7,<14",
  "httpx>=0.27,<1",
  "anyio>=4.4,<5",
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "jinja2>=3.1,<4",
  "matplotlib>=3.8,<4",
  "plotly>=5.22,<6",
  "jsonschema>=4.22,<5",
  "sqlalchemy>=2.0,<3",
  "orjson>=3.10,<4",
  "tqdm>=4.66,<5",
  "pymoo>=0.6.1,<0.8",
]

[project.optional-dependencies]
# CPU quickstart (optional)
cpu = [
  "transformers>=4.43,<5",
  "accelerate>=0.33,<1",
  # Torch is platform-specific. Prefer installing via conda for stability.
  # If you want pip: uncomment and choose CPU wheel explicitly for your platform.
  # "torch>=2.3,<3",
]

# GPU production run (optional)
gpu = [
  # vLLM is best installed on Linux + CUDA. Keep it optional.
  "vllm>=0.6,<1",
  # NVML bindings for GPU metrics (VRAM/util). Requires NVIDIA driver.
  "nvidia-ml-py>=12.555,<13",
]

# Development tooling (optional)
dev = [
  "pytest>=8.2,<9",
  "pytest-xdist>=3.6,<4",
  "pytest-cov>=5.0,<6",
  "ruff>=0.6.9,<1",
  "black>=24.8,<25",
  "mypy>=1.11,<2",
  "types-PyYAML>=6.0.12.20240808",
  "pip-tools>=7.4,<8",
  "pre-commit>=3.8,<4",
]

[project.scripts]
sbt = "servingbench_tuner.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/servingbench_tuner"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["results", "reports", "data"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # handled by formatter

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402", "F841", "SIM105"]
"scripts/run_formal_comparison_stability.py" = ["B023"]
